moved classes used only by BLOBs
