+++ src/main/java/io/chino/java/Blobs.java
@@ -3,8 +3,8 @@ package io.chino.java;
 import com.fasterxml.jackson.databind.JsonNode;
 import io.chino.api.blob.*;
 import io.chino.api.common.ChinoApiException;
-import io.chino.api.common.MD5Calc;
-import io.chino.api.common.SHA1Calc;
+import io.chino.api.blob.MD5Calc;
+import io.chino.api.blob.SHA1Calc;
 import io.chino.api.document.Document;
 import okhttp3.Request;
 import okhttp3.Response;
